Fixed bugs

+++ Lab1/Lab1/Circle.h
@@ -9,6 +9,6 @@ public:
 	~CCircle();
 
 private:
-	double m_radius;
+	int m_radius;
 };
